Vidi_7B/requirements.txt
transformers==4.44.2
sentencepiece==0.1.99
shortuuid
accelerate==0.33.0
peft
bitsandbytes
pydantic
numpy
scikit-learn==1.2.2
requests
httpx==0.22.0
decord
pillow
uvicorn
fastapi
einops==0.6.1
einops-exts==0.0.4
timm==0.6.13
deepspeed==0.14.5
ninja
wandb
pandas
matplotlib